🐛 fix: Fix test failures and CI issues
- Fix InMemoryFileSystem.exists() to return Result<boolean, Error> properly
- Fix async exists() calls in mkdir, writeTextFile, and remove methods
- Update command-helpers.ts to avoid env permission issues in tests
- Add mock config to test setup to prevent env var access during initialization
- Update path-integration test to handle OS-specific error messages

All tests now pass without requiring environment variable permissions.
