+++ test/commands/read.test.ts
@@ -35,6 +35,26 @@ describe("read command", () => {
       taskService: mockTaskService,
       gitService: mockGitService,
       i18nService: mockI18n,
+      config: {
+        task_directory: "~/locus",
+        git: {
+          extract_username: true,
+          username_from_remote: true,
+        },
+        file_naming: {
+          pattern: "{date}-{slug}-{hash}.md",
+          date_format: "YYYY-MM-DD",
+          hash_length: 8,
+        },
+        defaults: {
+          status: "todo",
+          priority: "normal",
+          tags: [],
+        },
+        language: {
+          default: "ja",
+        },
+      },
     });
 
     // Capture console output
